Refactored RoadPieces
